store: move package out of commands
