+++ snapcraft/commands/account.py
@@ -29,9 +29,7 @@ from craft_cli import BaseCommand, emit
 from craft_cli.errors import ArgumentParsingError
 from overrides import overrides
 
-from snapcraft import utils
-
-from . import store
+from snapcraft import store, utils
 
 if TYPE_CHECKING:
     import argparse
