+++ snapcraft/commands/manage.py
@@ -22,9 +22,7 @@ from typing import TYPE_CHECKING
 from craft_cli import BaseCommand, emit
 from overrides import overrides
 
-from snapcraft import errors, utils
-
-from . import store
+from snapcraft import errors, store, utils
 
 if TYPE_CHECKING:
     import argparse
