+++ snapcraft/store/_legacy_account.py
