+++ snapcraft/store/channel_map.py
