+++ snapcraft/store/client.py
