+++ snapcraft/store/constants.py
