+++ tests/unit/conftest.py
@@ -14,11 +14,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import base64
+import textwrap
 from pathlib import Path
 from typing import Any, Dict, Optional, Tuple
 
 import pytest
 import yaml
+from pymacaroons import Caveat, Macaroon
 
 from snapcraft.extensions import extension, register, unregister
 
@@ -228,3 +231,76 @@ def fake_extension_name_from_legacy():
             return {"fake-extension-extra/fake-part": {"plugin": "nil"}}
 
     yield ExtensionImpl
+
+
+@pytest.fixture
+def fake_client(mocker):
+    """Forces get_client to return a fake craft_store.BaseClient"""
+    client = mocker.patch("craft_store.BaseClient", autospec=True)
+    mocker.patch("snapcraft.store.client.get_client", return_value=client)
+    return client
+
+
+@pytest.fixture
+def fake_confirmation_prompt(mocker):
+    """Fake the confirmation prompt."""
+    return mocker.patch(
+        "snapcraft.utils.confirm_with_user", return_value=False, autospec=True
+    )
+
+
+@pytest.fixture
+def root_macaroon():
+    return Macaroon(
+        location="fake-server.com",
+        signature="d9533461d7835e4851c7e3b639144406cf768597dea6e133232fbd2385a5c050",
+        caveats=[
+            Caveat(
+                caveat_id="1234567890",
+                location="fake-sso.com",
+                verification_key_id="1234567890",
+            )
+        ],
+    ).serialize()
+
+
+@pytest.fixture
+def discharged_macaroon():
+    return Macaroon(
+        location="fake-server.com",
+        signature="d9533461d7835e4851c7e3b639122406cf768597dea6e133232fbd2385a5c050",
+    ).serialize()
+
+
+@pytest.fixture(params=["encode", "no-encode"])
+def legacy_config_credentials(request):
+    config = textwrap.dedent(
+        f"""\
+        [login.ubuntu.com]
+        macaroon={root_macaroon}
+        unbound_discharge={discharged_macaroon}
+        """
+    )
+
+    if request.param == "encode":
+        return base64.b64encode(config.encode()).decode()
+
+    if request.param == "no-encode":
+        return config
+
+    raise RuntimeError("unhandled param")
+
+
+@pytest.fixture
+def legacy_config_path(
+    monkeypatch, new_dir, root_macaroon, discharged_macaroon, legacy_config_credentials
+):
+    config_file = new_dir / "snapcraft.cfg"
+    monkeypatch.setattr(
+        "snapcraft.store._legacy_account.LegacyUbuntuOne.CONFIG_PATH",
+        config_file,
+    )
+
+    config_file.write_text(legacy_config_credentials)
+
+    return config_file
