+++ tests/unit/store/utils.py
